drm: xlnx: fb: Use minimum pitch always
Previously, the function adjusts the pitch only when the given pitch
is not aligned with hardware requirement. Otherwise, the requested
pitch is used as is. This can give very wide range of pitches based
on the requested pitch value, so make it more consistent by returning
the minimum pitch always.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>